requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
argon2-cffi==21.1.0
async-timeout==4.0.1
attrs==21.2.0
avalanche-lib==0.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
botorch==0.4.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.2
cloudpickle==2.0.0
cycler==0.11.0
databricks-cli==0.16.6
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docker==5.0.3
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.6.0
Flask==2.1.1
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
gdown==4.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.3
google-auth-oauthlib==0.4.6
GPUtil==1.4.0
gpytorch==1.5.1
greenlet==1.1.2
grpcio==1.42.0
gunicorn==20.1.0
idna==3.3
ignite==1.1.0
imageio==2.13.1
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
mistune==0.8.4
mlflow==1.25.1
multidict==5.2.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numpy==1.20.1
oauthlib==3.1.1
opencv-python==4.5.1.48
openslide-python==1.1.2
packaging==21.3
pandas==1.2.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prometheus-flask-exporter==0.20.1
promise==2.3
prompt-toolkit==3.0.23
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pyDeprecate==0.3.1
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.6
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytorch-lightning @ https://github.com/PyTorchLightning/pytorch-lightning/archive/master.zip
pytorchcv==0.0.67
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
qtconsole==5.2.1
QtPy==1.11.3
quadprog==0.1.11
querystring-parser==1.2.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.8.0
sentry-sdk==1.5.10
setproctitle==1.2.3
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
sox==1.4.1
SQLAlchemy==1.4.35
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchmetrics==0.6.1
torchtext==0.9.1
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
wandb==0.12.14
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0
